Revert "Define profile in Procfile."
This reverts commit bde19a78c89b8034d796f5ef241caabe1bf1bb78.
Spring Boot / JPA / Hibernate need more configuration for this
to work. Fails on not figuring out the hibernate dialect,
probably more problems will surface after that. Also, need to
address creating and updating the PostgreSQL DB, since it is
not automatically synced with code (of course could for the
time being enable that?).