Validation to happen in controllers.
Yes, it's kind of handy to do it in the service, where it's
easier to avoid duplication of code. But passing validation errors
as exception messages is just *wrong*, and if we should do i18n
of the messages, we would have to "leak" user locale to the
services, etc... And validation of input should happen close to where
input is received, ie. in the controller. It's a matter of principle.